对容器的简单支持，以及简单美观一下代码

+++ src/main/java/samples/ServiceBean.java
@@ -2,13 +2,13 @@ package samples;
 
 public class ServiceBean extends SuperSample {
 
-    private int               intV;
+    private long              intV;
     private String            stringV;
     private ServiceBeanHelper helper;
     private Saver             saver;
-    protected int  protectedI;
+    protected int             protectedI;
 
-    public boolean shouldHelpNow(int type) {
+    public boolean shouldHelpNow(long type) {
         if (type > intV) {
             return false;
         }
